se reorganizaron los componentes utilizados en el ambito del lobby
