+++ src/components/room/lobby/Results.vue
@@ -6,13 +6,13 @@
             <TabChat />
         </div>
 
+        <VoteState />
         <button class="btn btn-primary" v-if="room.privileges && !socketStore.currentProposal.last_prop"
             @click="nextProposal">
             Siguiente propuesta
         </button>
 
         <h1 v-if="socketStore.currentProposal.last_prop">fin de votacion</h1>
-
         <!-- <VotesScroll /> -->
     </div>
 </template>
@@ -20,10 +20,11 @@
 <script setup>
 import { provide, onMounted, ref, computed, watch } from 'vue';
 import { useWebSocketStore } from '../stores/socketStore'
-import TabChat from './TabChat.vue';
+import TabChat from './components/TabChat.vue';
 import VotesScroll from './Results.vue';
+import VoteState from './components/VoteState.vue';
 import { useRouter } from 'vue-router';
-import CardResult from './CardResult.vue'
+import CardResult from './components/CardResult.vue'
 
 const socketStore = useWebSocketStore();
