+++ src/pages/protected/roomInfo/[id].astro
@@ -2,7 +2,7 @@
 import RoomInfo from "../../../components/room/RoomInfo.vue";
 import ProtectedRoute from "../../../components/authentication/ProtectedRoute.vue";
 import Layout from "../../../layouts/Layout.astro";
-import StartRoom from "../../../components/room/lobby/StartRoom.vue";
+import StartRoom from "../../../components/room/lobby/components/StartRoom.vue";
 
 const { id } = Astro.params;
 ---
